Update webkit1gtk3 to 2.4.9
This release contains fixes for two of the patches that we carry, so
those can be removed. One of the remaining patches (webkit-140167-
disable-netscape-api) is unlikely to be fixed, and the other one (webkit-
133293-cfi-clang-failure) was actually a clang bug that has been fixed in
XCode 6.3. We'll leave it in since it doesn't seem to do any harm, and
XCode 6.3 might not be present on older systems.