WebKit and dependencies in new modules file
Having WebKit in "unsupported" is kind of misleading now that it's
getting easier to build WebKit. It's also required for a lot of modern
Gnome applications.

Let's move WebKit and its networking / crypto stack dependencies into a
new gtk-osx-webkit.modules file. It's technically still "unsupported"
according to the comment in gtk-osx.modules, just a little... less so.